feat: generate budget with total

+++ src/main/java/org/sistema/arroz/riceservice/modules/supplyFormula/application/port/in/GetSuppliesBudgetUseCase.java
@@ -4,5 +4,5 @@ package org.sistema.arroz.riceservice.modules.supplyFormula.application.port.in;
 import java.util.List;
 
 public interface GetSuppliesBudgetUseCase {
-    List<GetSuppliesBudgetDTO> getSuppliesBudget(Long productId, Integer hectares);
+    List<GetSuppliesBudgetDTO> getSuppliesBudget(Long productId, Double hectares);
 }
